Remove unused widget and add youtubeview in GameActivity
